[hail] Remove `LiftNonCompilable` compiler pass
Should be unnecessary now.

Stacked on #7421
