+++ hail/src/main/scala/is/hail/expr/ir/InterpretNonCompilable.scala
@@ -3,12 +3,30 @@ package is.hail.expr.ir
 import is.hail.expr.types.virtual.TStruct
 import org.apache.spark.sql.Row
 
+import scala.collection.mutable
+
 object InterpretNonCompilable {
+
+  def extractNonCompilable(irs: IR*): Map[String, IR] = {
+    val included = mutable.Set.empty[IR]
+
+    def visit(ir: IR): Unit = {
+      if (!Compilable(ir) && !included.contains(ir))
+        included += ir
+      ir.children.foreach {
+        case ir: IR => visit(ir)
+        case _ =>
+      }
+    }
+
+    irs.foreach(visit)
+    included.toArray.map { l => genUID() -> l }.toMap
+  }
 
   def apply(ctx: ExecuteContext, ir: IR): (IR, Row, TStruct, String) = {
 
     val name = genUID()
-    val nonCompilableNodes = LiftNonCompilable.extractNonCompilable(ir).toArray
+    val nonCompilableNodes = extractNonCompilable(ir).toArray
     val emittable = nonCompilableNodes.filter { case (_, value) => CanEmit(value.typ) }
     val nonEmittable = nonCompilableNodes.filter { case (_, value) => !CanEmit(value.typ) }
